Change directory structure for gem
