+++ aws_iot.gemspec
@@ -6,8 +6,8 @@ require 'aws_iot/version'
 Gem::Specification.new do |spec|
   spec.name          = "aws_iot"
   spec.version       = AwsIot::VERSION
-  spec.authors       = ["Hiroyuki Sato"]
-  spec.email         = ["h-sato@ruby-dev.jp"]
+  spec.authors       = ["Pierre Goudet"]
+  spec.email         = ["p-goudet@ruby-dev.jp"]
 
   spec.summary       = %q{TODO: Write a short summary, because Rubygems requires one.}
   spec.description   = %q{TODO: Write a longer description or delete this line.}
@@ -28,6 +28,12 @@ Gem::Specification.new do |spec|
   spec.require_paths = ["lib"]
 
   spec.add_development_dependency "bundler", "~> 1.10"
+  spec.add_development_dependency "pry"
   spec.add_development_dependency "rake", "~> 10.0"
   spec.add_development_dependency "rspec"
+
+  spec.add_runtime_dependency "facets", "~> 3.1.0"
+  spec.add_runtime_dependency "json", "~> 1.8.3"
+  spec.add_runtime_dependency "mqtt", "~> 0.4.0"
+  spec.add_runtime_dependency "timers", "~> 4.1.1"
 end
