+++ lib/aws_iot.rb
@@ -1,4 +1,6 @@
 require "aws_iot/version"
+require "aws_iot/mqtt_adapter"
+require "aws_iot/mqtt_shadow_client"
 
 module AwsIot
   # Your code goes here...
