+++ lib/aws_iot/mqtt_shadow_client.rb
@@ -1,0 +1,6 @@
+require 'aws_iot/mqtt_shadow_client/shadow_client'
+
+module AwsIot
+  module MqttShadowClient
+  end
+end
