+++ lib/aws_iot/mqtt_shadow_client/json_payload_parser.rb
@@ -1,0 +1,34 @@
+require 'securerandom'
+require 'timers'
+require 'thread'
+require 'json'
+
+module AwsIot
+  module MqttShadowClient
+    class JSONPayloadParser
+      ### This class acts as Basic JSON parser.
+      ### The answer from AWS is in a JSON format.
+      ### All different key of the JSON file should be defined as hash key
+
+      def initialize
+        @message = {}
+      end
+
+      def set_message(message)
+        @message = JSON.parse(message)
+      end
+
+      def get_attribute_value(key)
+        @message[key]
+      end
+
+      def set_attribute_value(key, value)
+        @message[key] = value
+      end
+
+      def get_json
+        @message.to_json
+      end
+    end
+  end
+end
